pyproject.toml
[project]
name = "finsight"
version = "0.1.0"
description = "FastAPI-based REST API for integrating with Tinkoff Invest and predicting short-term stock movements."
authors = [
  { name = "Aleksander Moiseenko", email = "alex.mois.developer@gmail.com" }
]
license = "MIT"
readme = "README.md"
keywords = ["fastapi", "ml", "finance", "stock", "predictions", "tinkoff"]
requires-python = ">=3.13"

classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.13",
  "Framework :: FastAPI"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<4"
bcrypt = "==4.0.1"
fastapi = "==0.115.11"
pydantic = {extras = ["email"], version = "==2.10.6"}
pydantic-settings = "==2.8.1"
structlog = "==25.2.0"
uvicorn = "==0.30.0"
orjson = "==3.10.16"
dependency-injector = "==4.46.0"
tinkoff-investments = "==0.2.0b110"
click = "==8.1.8"

[tool.poetry.group.dev.dependencies]
pre-commit = "==4.1.0"
ruff = "==0.9.6"
mypy = "==1.15.0"
commitizen = "==4.2.1"
grpc-stubs = "==1.53.0.5"

[tool.poetry.group.test.dependencies]
httpx = "==0.28.1"
pytest = "==8.3.4"
pytest-asyncio = "==0.25.3"
pytest-cov = "==6.0.0"
pytest-mock = "==3.14.0"
